Support specifying database
Redix/Redix support specifying a number for a database to connect to which allows you to partition
your keys into different databases. This allows you to, for example, flush the keys for a sub-section
of your keys without effecting other keys. This change passes through the database provided in the config
if one is given and defaults to nil, which means Redix will use the default Redis database 0