Add extend functionality to Redlock
Add the ability for the client to extend a lock's TTL. This is useful in situations where a client needs to lock a resource for a long time but still wants a relatively shorter TTL in case the lock-holder crashes or exits unexpectedly. The lock-holder can continually extend its ownership of the lock while it still needs it.

I tested the new functionality manually and added unit tests to further test this functionality.