Allow specification of log levels
The redlock library can be pretty aggressive with its logging. Sometimes users of the library do not need as much logging and may only want to be notified on important failures. For example, failure to acquire a lock may not be considered worth logging, but a disconnection from Redis would be.

The existing code has a hidden flag called "show_debug_logs" that be used to control logging a little bit. This PR changes the existing flag to a "log_level" config that allows the user to specify "debug", "info", "warning", or "error" level logs. This offers much greater control of redlock's logging. I'm keeping backwards compatibility by allowing the old "show_debug_logs" flag to override the "log_level" to "debug".

Additional changes include:
- Failure to acquire a lock eventually is now considered info rather than warning
- Changed `Logger.warn` to `Logger.warning` because `Logger.warn` is deprecated in the Logger library